ai_agent.1.2/requirements.txt
requirements.txt
llama-cpp-python>=0.2.0
pyautogui>=0.9.53
pillow>=9.0.0
pytesseract>=0.3.10
rich>=13.0.0
